Use single file module
