--- crates/runtime/src/dataconnector/graphql/table_provider.rs
@@ -1,78 +1,0 @@
-/*
-Copyright 2024 The Spice.ai OSS Authors
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-     https://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-use std::{any::Any, sync::Arc};
-
-use arrow::datatypes::SchemaRef;
-use async_trait::async_trait;
-use data_components::arrow::write::MemTable;
-use datafusion::{
-    datasource::{TableProvider, TableType},
-    error::DataFusionError,
-    execution::context::SessionState,
-    logical_expr::Expr,
-    physical_plan::ExecutionPlan,
-};
-use futures::TryFutureExt;
-
-use super::client::GraphQLClient;
-
-#[allow(clippy::module_name_repetitions)]
-pub struct GraphQLTableProvider {
-    client: GraphQLClient,
-    schema: SchemaRef,
-}
-
-impl GraphQLTableProvider {
-    pub async fn new(client: GraphQLClient) -> super::client::Result<Self> {
-        let (_, schema, _) = client.execute(None, None, None).await?;
-
-        Ok(Self { client, schema })
-    }
-}
-
-#[async_trait]
-impl TableProvider for GraphQLTableProvider {
-    fn as_any(&self) -> &dyn Any {
-        self
-    }
-
-    fn schema(&self) -> SchemaRef {
-        Arc::clone(&self.schema)
-    }
-
-    fn table_type(&self) -> TableType {
-        TableType::Base
-    }
-
-    async fn scan(
-        &self,
-        state: &SessionState,
-        projection: Option<&Vec<usize>>,
-        filters: &[Expr],
-        limit: Option<usize>,
-    ) -> datafusion::error::Result<Arc<dyn ExecutionPlan>> {
-        let res = self
-            .client
-            .execute_paginated(Arc::clone(&self.schema), limit)
-            .map_err(|e| DataFusionError::Execution(format!("{e}")))
-            .await?;
-
-        let table = MemTable::try_new(Arc::clone(&self.schema), res)?;
-
-        table.scan(state, projection, filters, limit).await
-    }
-}
